pyproject.toml
[project]
name = "mcpllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chardet>=5.2.0",
    "chromadb>=1.3.0",
    "coverage>=7.3.0",
    "fastapi>=0.121.2",
    "flashrank>=0.2.10",
    "html5lib>=1.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=0.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=0.1.0",
    "langchain-google-genai>=3.0.2",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langgraph>=0.0.40",
    "lxml>=6.0.2",
    "markdown2>=2.5.4",
    "mcp>=1.19.0",
    "notebook>=7.4.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "plotly>=6.3.1",
    "pydantic>=2.12.3",
    "pydantic-core>=2.41.4",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "sentence-transformers>=5.1.2",
    "splunk-sdk>=2.1.1",
    "streamlit>=1.51.0",
    "unstructured[all-docs]>=0.18.18",
    "uvicorn[standard]>=0.38.0",
]
